// src/routes/darkpearl/types.ts
export type GitHubAuth = {
	access_token: string
	user_id: number
	username: string
	avatar_url: string
	connected_at: string
}